guardianship/requirements.txt
chromadb==0.5.0
langchain==0.2.12
langchain_anthropic==0.1.22
langchain_community==0.2.11
langchain_core==0.2.28
langgraph==0.1.19
openpyxl==3.1.5
pydantic==2.8.2
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
pywin32==306
Requests==2.32.3
